Fix native setup for FirebaseAdmin
Usage of FirebaseAdmin fails in a native environment. These changes should fix that. Note that not all uses have been validated, so we may need other changes than these, but its a start.
