+++ .github/workflows/build.yml
@@ -45,5 +45,5 @@ jobs:
 
     - name: Build in native
       run: |
-        ./mvnw -B -Pnative package --file integration-tests/main/pom.xml
+        ./mvnw -B -Pnative verify --file integration-tests/pom.xml
