+++ integration-tests/firebase-admin/src/test/java/io/quarkiverse/googlecloudservices/it/firebaseadmin/FirebaseAuthResourceTest.java
@@ -2,10 +2,7 @@ package io.quarkiverse.googlecloudservices.it.firebaseadmin;
 
 import static io.restassured.RestAssured.given;
 import static org.hamcrest.Matchers.anEmptyMap;
-import static org.hamcrest.Matchers.blankOrNullString;
 import static org.hamcrest.Matchers.equalTo;
-import static org.hamcrest.Matchers.hasEntry;
-import static org.hamcrest.Matchers.not;
 
 import java.util.Map;
 import java.util.Set;
@@ -21,55 +18,10 @@ import io.quarkus.test.junit.QuarkusTest;
 import io.restassured.http.ContentType;
 
 @QuarkusTest
-class FirebaseAuthResourceTest extends FirebaseAuthTest {
+class FirebaseAuthResourceTest extends FirebaseAuthUserResourceTest {
 
     @Inject
     FirebaseAuth firebaseAuth;
-
-    @Test
-    void shouldCreateUser() {
-        given()
-                .contentType(ContentType.JSON)
-                .queryParam("email", "johndoe@quarkusio.com")
-                .queryParam("displayName", "John Doe")
-                .post("/auth/users/create")
-                .then()
-                .log().ifValidationFails()
-                .statusCode(200)
-                .body("uid", not(blankOrNullString()))
-                .body("email", equalTo("johndoe@quarkusio.com"))
-                .body("displayName", equalTo("John Doe"));
-    }
-
-    @Test
-    void shouldStoreCustomUserClaims() {
-        given()
-                .contentType(ContentType.JSON)
-                .queryParam("uid", "12345")
-                .queryParam("email", "johndoe@quarkusio.com")
-                .queryParam("displayName", "John Doe")
-                .post("/auth/users/create")
-                .then()
-                .log().ifValidationFails()
-                .statusCode(200)
-                .body("uid", equalTo("12345"))
-                .body("customClaims", anEmptyMap());
-
-        given()
-                .contentType(ContentType.JSON)
-                .body(Map.of("role", "admin"))
-                .put("/auth/users/{uid}/claims", 12345)
-                .then()
-                .log().ifValidationFails()
-                .statusCode(204);
-
-        given()
-                .get("/auth/users/{uid}", 12345)
-                .then()
-                .log().ifValidationFails()
-                .statusCode(200)
-                .body("customClaims", hasEntry("role", "admin"));
-    }
 
     @Test
     void shouldHandleRoles() throws FirebaseAuthException {
