+++ integration-tests/firebase-admin/src/test/java/io/quarkiverse/googlecloudservices/it/firebaseadmin/FirebaseAuthTest.java
@@ -2,19 +2,27 @@ package io.quarkiverse.googlecloudservices.it.firebaseadmin;
 
 import static io.restassured.RestAssured.given;
 
-import org.eclipse.microprofile.config.inject.ConfigProperty;
+import org.eclipse.microprofile.config.ConfigProvider;
 import org.junit.jupiter.api.BeforeEach;
 
+/**
+ * Base class for authorization tests.
+ * <p>
+ * {@link ConfigProvider} is used to retreive the config valeus over injection as this also needs to run in a integration
+ * test.
+ */
 public abstract class FirebaseAuthTest {
 
-    @ConfigProperty(name = "quarkus.google.cloud.project-id")
     String projectId;
 
-    @ConfigProperty(name = "quarkus.google.cloud.firebase.auth.emulator-host")
     String emulatorHost;
 
     @BeforeEach
     public void deleteAllAccounts() {
+        projectId = ConfigProvider.getConfig().getValue("quarkus.google.cloud.project-id", String.class);
+        emulatorHost = ConfigProvider.getConfig().getValue("quarkus.google.cloud.firebase.auth.emulator-host",
+                String.class);
+
         var emulatorHostParts = emulatorHost.split(":");
         var port = emulatorHostParts.length == 2 ? Integer.parseInt(emulatorHostParts[1]) : 9099;
